+++ src/App.vue
@@ -1,7 +1,7 @@
 <template>
   <div>
-    <cookie-notice>
-      <template v-slot:content>
+    <cookie-notice :fields="cookieFields">
+      <template slot="content">
         <h5>Cookies</h5>
         <p>
           We use <a href="#">cookies</a> in order to improve the user experience Maecenas fauci
@@ -11,12 +11,30 @@
           parturient montes, nascetur ridiculus mus.
         </p>
       </template>
+      <template slot="settings">
+        Settings
+      </template>
     </cookie-notice>
+    <h2 class="js-toggle-cookie-settings">
+      Toggle cookie notice
+    </h2>
   </div>
 </template>
 
 <script>
 export default {
-  //
+  data() {
+    return {
+      cookieFields: [
+        {
+          id: 'necessary',
+          label: 'Noodzakelijk',
+          disabled: true,
+          checked: true,
+        },
+        { id: 'statistics', label: 'Statistieken', checked: false },
+      ],
+    };
+  },
 };
 </script>