+++ src/components/cookie-notice/event.js
@@ -1,0 +1,7 @@
+/**
+ * Global event for CookieNotice
+ */
+
+import Vue from 'vue'; // eslint-disable-line import/no-extraneous-dependencies
+
+export const CookieEvent = new Vue(); // eslint-disable-line import/prefer-default-export