+++ vue.config.js
@@ -1,8 +1,8 @@
 module.exports = {
   css: {
     loaderOptions: {
-      scss: {
-        prependData: '@import "~@/scss/_config.scss";',
+      sass: {
+        additionalData: '@import "~@/scss/_config.scss";',
       },
     },
   },