+++ CHANGELOG.md
@@ -6,10 +6,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2-beta] - 2020-07-21
+### Added
+- `CookieNotice` is now GDPR ready
+- You can now customize checkboxes and toggle settings
+- Added global `CookieEvent`
+
 ## [0.0.1-beta] - 2020-04-16
 ### Added
 - First beta version of the Vue component library
 - Added `CookieNotice` component
 
-[Unreleased]: https://github.com/get-uncommon/vue-components/compare/v0.0.1-beta...HEAD
+[Unreleased]: https://github.com/get-uncommon/vue-components/compare/v0.0.2-beta...HEAD
 [0.0.1-beta]: https://github.com/get-uncommon/vue-components/releases/tag/v0.0.1-beta
+[0.0.2-beta]: https://github.com/get-uncommon/vue-components/releases/tag/v0.0.2-beta